% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank.R
\name{download_genbank}
\alias{download_genbank}
\title{download_genbank}
\usage{
download_genbank(accession)
}
\arguments{
\item{accession}{accession number}
}
\description{
download genbank files
}
\author{
Guangchuang Yu
}
