% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{global_aln}
\alias{global_aln}
\title{global_aln}
\usage{
global_aln(X, Y, score = list(match = 5, mismatch = -2, indel = -6))
}
\arguments{
\item{X}{sequence 1}

\item{Y}{sequence 2}

\item{score}{score system}
}
\value{
smuAlign object
}
\description{
global alignment
}
\references{
\url{https://mp.weixin.qq.com/s/4DL-pJVItOhkYJAbPEg7BQ}
}
\author{
Guangchuang Yu
}
