% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbank.R
\name{read.genbank}
\alias{read.genbank}
\title{read.genbank}
\usage{
read.genbank(file)
}
\arguments{
\item{file}{input genbank file}
}
\value{
sequence object
}
\description{
extract accession number and sequence from genbank file
}
\author{
Guangchuang Yu
}
