% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{write.fasta}
\alias{write.fasta}
\title{write.fasta}
\usage{
write.fasta(x, file)
}
\arguments{
\item{x}{DNA or protein sequence object}

\item{file}{output file}
}
\description{
write DNA or protein sequences to FASTA file
}
